fix: remove git from skills
Oh boy! Git is a whole universe in itself, cannot really rate myself on
how much of Git I know unless I explore the rest of it. So to avoid having
to give a rating, better not to mention it. People can obviously see that
I have hosted it on GitHub and that I maintain all my projects here so
they will come to know that I know to use Git/GitHub.